+++ src/component/pages/Map/Map.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import mapboxgl from 'mapbox-gl';
+import {token} from 'constants.js';
 
-mapboxgl.accessToken = 'pk.eyJ1Ijoic2FzaGEtZ3JhdmVzIiwiYSI6ImNrZjVlY3dvMTBtOXQyem1kd2kxeGlha2YifQ.4h4RYOVB0aR8aUH5c2eJcg';
-
+mapboxgl.accessToken = token;
 
 class Map extends React.Component {
