+++ src/constants.js
@@ -1,0 +1,1 @@
+export const token = 'pk.eyJ1Ijoic2FzaGEtZ3JhdmVzIiwiYSI6ImNrZjVlY3dvMTBtOXQyem1kd2kxeGlha2YifQ.4h4RYOVB0aR8aUH5c2eJcg';
